+++ lib/membrane/bin/callback_context/setup.ex
@@ -1,0 +1,6 @@
+defmodule Membrane.Bin.CallbackContext.Setup do
+  @moduledoc """
+  Structure representing a context that is passed to the `c:Membrane.Bin.handle_setup/2` callback.
+  """
+  use Membrane.Core.Bin.CallbackContext
+end